+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Mapbox welcomes participation and contributions from everyone.
 ## 11.22.0-rc.1 - 08 April, 2026
 
 * Fix old 3D puck model remaining after updating the puck model.
+## 11.21.1 - 10 April, 2026
 
 ## 11.21.0 - 02 April, 2026